+++ pakery-tests/tests/opaque_vectors.rs
@@ -101,11 +101,6 @@ impl rand_core::RngCore for SequentialRng {
             written += to_copy;
         }
     }
-
-    fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), rand_core::Error> {
-        self.fill_bytes(dest);
-        Ok(())
-    }
 }
 
 impl rand_core::CryptoRng for SequentialRng {}
@@ -674,7 +669,7 @@ fn test_vector2_full_login() {
 
 #[test]
 fn test_full_roundtrip_random() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
     let password = b"correct horse battery staple";
 
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
@@ -715,7 +710,7 @@ fn test_full_roundtrip_random() {
 
 #[test]
 fn test_wrong_password() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
 
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
 
@@ -742,7 +737,7 @@ fn test_wrong_password() {
 
 #[test]
 fn test_tampered_server_mac() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
 
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
 
@@ -772,7 +767,7 @@ fn test_tampered_server_mac() {
 
 #[test]
 fn test_tampered_client_mac() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
 
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
 
@@ -810,7 +805,7 @@ fn test_tampered_client_mac() {
 fn test_fake_credential_response() {
     use pakery_opaque::messages::CredentialResponse;
 
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
     let password = b"some password";
 
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
@@ -840,7 +835,7 @@ fn test_fake_credential_response() {
 
 #[test]
 fn test_fake_credential_client_fails() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
     let password = b"some password";
 
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
@@ -896,7 +891,7 @@ mod argon2_tests {
 
     #[test]
     fn test_argon2_roundtrip() {
-        let mut rng = rand_core::OsRng;
+        let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
         let password = b"correct horse battery staple";
 
         let setup = ServerSetup::<OpaqueRistretto255Sha512Argon2>::new(&mut rng).unwrap();
@@ -941,7 +936,7 @@ mod argon2_tests {
 
     #[test]
     fn test_argon2_wrong_password() {
-        let mut rng = rand_core::OsRng;
+        let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
 
         let setup = ServerSetup::<OpaqueRistretto255Sha512Argon2>::new(&mut rng).unwrap();
 
@@ -977,7 +972,7 @@ mod argon2_tests {
 
 #[test]
 fn test_start_fake_ke2_size_matches_real() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
     let password = b"test password";
 
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
@@ -1068,7 +1063,7 @@ fn test_start_fake_ke2_size_matches_real() {
 
 #[test]
 fn test_empty_password_roundtrip() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
     let password = b"";
 
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
@@ -1249,7 +1244,7 @@ fn test_deterministic_replay() {
 
 #[test]
 fn test_tampered_ke2_evaluated_message() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
 
     let (reg_request, reg_state) =
@@ -1278,7 +1273,7 @@ fn test_tampered_ke2_evaluated_message() {
 
 #[test]
 fn test_tampered_ke2_masked_response() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
 
     let (reg_request, reg_state) =
@@ -1307,7 +1302,7 @@ fn test_tampered_ke2_masked_response() {
 
 #[test]
 fn test_tampered_ke2_server_keyshare() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
 
     let (reg_request, reg_state) =
@@ -1336,7 +1331,7 @@ fn test_tampered_ke2_server_keyshare() {
 
 #[test]
 fn test_context_mismatch_fails() {
-    let mut rng = rand_core::OsRng;
+    let mut rng = rand_core::UnwrapErr(rand_core::OsRng);
     let setup = ServerSetup::<OpaqueRistretto255Sha512>::new(&mut rng).unwrap();
 
     let (reg_request, reg_state) =